@@ -1,10 +1,11 @@
-﻿using Dicom;
-using Dicom.Imaging;
-using Dicom.IO.Buffer;
+﻿using FellowOakDicom;
+using FellowOakDicom.IO.Buffer;
 using System;
+using System.Collections.Generic;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.Runtime.InteropServices;
+using FellowOakDicom.Imaging;
 
 namespace BadMedicine.Dicom
 {
@@ -13,36 +14,34 @@ namespace BadMedicine.Dicom
     /// </summary>
     internal class PixelDrawer
     {
-        readonly SolidBrush _blackBrush = new SolidBrush(Color.Black);
-        readonly SolidBrush _whiteBrush = new SolidBrush(Color.White);
+        readonly SolidBrush _blackBrush = new(Color.Black);
+        readonly SolidBrush _whiteBrush = new(Color.White);
 
         internal void DrawBlackBoxWithWhiteText(DicomDataset ds, int width, int height, string msg)
         {
-            using (var bitmap = new Bitmap(500, 500))
+            using var bitmap = new Bitmap(500, 500);
+            using (var g = Graphics.FromImage(bitmap))
             {
-                using (var g = Graphics.FromImage(bitmap))
-                {
-                    g.FillRectangle(_blackBrush, 0, 0, width, height);
-                    using (var font = new Font(FontFamily.GenericMonospace, 12))
-                        g.DrawString(msg, font, _whiteBrush, 250, 100);
-                }
-
-                byte[] pixels = GetPixels(bitmap, out int rows, out int columns);
-                MemoryByteBuffer buffer = new MemoryByteBuffer(pixels);
-
-                ds.Add(DicomTag.PhotometricInterpretation, PhotometricInterpretation.Rgb.Value);
-                ds.Add(DicomTag.Rows, (ushort)rows);
-                ds.Add(DicomTag.Columns, (ushort)columns);
-                ds.Add(DicomTag.BitsAllocated, (ushort)8);
-
-                DicomPixelData pixelData = DicomPixelData.Create(ds, true);
-                pixelData.BitsStored = 8;
-                pixelData.SamplesPerPixel = 3;
-                pixelData.HighBit = 7;
-                pixelData.PixelRepresentation = 0;
-                pixelData.PlanarConfiguration = 0;
-                pixelData.AddFrame(buffer);
+                g.FillRectangle(_blackBrush, 0, 0, width, height);
+                using var font = new Font(FontFamily.GenericMonospace, 12);
+                g.DrawString(msg, font, _whiteBrush, 250, 100);
             }
+
+            byte[] pixels = GetPixels(bitmap, out int rows, out int columns);
+            MemoryByteBuffer buffer = new(pixels);
+
+            ds.Add(DicomTag.PhotometricInterpretation, PhotometricInterpretation.Rgb.Value);
+            ds.Add(DicomTag.Rows, (ushort)rows);
+            ds.Add(DicomTag.Columns, (ushort)columns);
+            ds.Add(DicomTag.BitsAllocated, (ushort)8);
+
+            DicomPixelData pixelData = DicomPixelData.Create(ds, true);
+            pixelData.BitsStored = 8;
+            pixelData.SamplesPerPixel = 3;
+            pixelData.HighBit = 7;
+            pixelData.PixelRepresentation = 0;
+            pixelData.PlanarConfiguration = 0;
+            pixelData.AddFrame(buffer);
         }
 
         private static byte[] GetPixels(Bitmap image, out int rows, out int columns)
@@ -53,7 +52,7 @@ namespace BadMedicine.Dicom
             if (rows % 2 != 0 && columns % 2 != 0)
                 --columns;
 
-            BitmapData data = image.LockBits(new Rectangle(0, 0, columns, rows), ImageLockMode.ReadOnly, image.PixelFormat);
+            BitmapData data = image.LockBits(new(0, 0, columns, rows), ImageLockMode.ReadOnly, image.PixelFormat);
             IntPtr bmpData = data.Scan0;
             try
             {
@@ -61,7 +60,7 @@ namespace BadMedicine.Dicom
                 int size = rows * stride;
                 byte[] pixelData = new byte[size];
                 for (int i = 0; i < rows; ++i)
-                    Marshal.Copy(new IntPtr(bmpData.ToInt64() + i * data.Stride), pixelData, i * stride, stride);
+                    Marshal.Copy(new(bmpData.ToInt64() + i * data.Stride), pixelData, i * stride, stride);
 
                 //swap BGR to RGB
                 SwapRedBlue(pixelData);
@@ -72,13 +71,11 @@ namespace BadMedicine.Dicom
                 image.UnlockBits(data);
             }
         }
-        private static void SwapRedBlue(byte[] pixels)
+        private static void SwapRedBlue(IList<byte> pixels)
         {
-            for (int i = 0; i < pixels.Length; i += 3)
+            for (var i = 0; i < pixels.Count; i += 3)
             {
-                byte temp = pixels[i];
-                pixels[i] = pixels[i + 2];
-                pixels[i + 2] = temp;
+                (pixels[i], pixels[i + 2]) = (pixels[i + 2], pixels[i]);
             }
         }
     }
